commit tree 8fd6893a523f714ac9348e332c12db26ef9b0aa0

+++ app/controllers/spices_controller.rb
@@ -1,7 +1,7 @@
 class SpicesController < ApplicationController
   def index
-    spices = Spice.all
-    render json: spices
+    spice = Spice.all
+    render json: spice
   end
 
   def create
@@ -10,22 +10,26 @@ class SpicesController < ApplicationController
   end
 
   def update
-    spice = find_spice
-    spice.update(spice_params)
-    render json: spice
+    spice = Spice.find_by(id: params[:id])
+    if spice
+      spice.update(spice_params)
+      render json: spice
+    else
+      render json: { error: "Spice not found" }, status: :not_found
+    end
   end
 
   def destroy
-    spice = find_spice
-    spice.destroy
-    head :no_content
+    spice = Spice.find_by(id: params[:id])
+    if spice
+      spice.destroy
+      head :no_content
+    else
+      render json: { error: "Spice not found" }, status: :not_found
+    end
   end
 
   private
-
-  def find_spice
-    Spice.find(params[:id])
-  end
 
   def spice_params
     params.permit(:title, :image, :description, :notes, :rating)
